Move describe.py to Describer/Describer.py and main_describe.py to describe.py
